+++ lib/td_ai_web/controllers/suggestion_controller.ex
@@ -3,6 +3,8 @@ defmodule TdAiWeb.SuggestionController do
 
   alias TdAi.Completion
   alias TdAi.Completion.Suggestion
+  alias TdAi.FieldCompletion
+  alias TdDfLib.Templates
 
   action_fallback TdAiWeb.FallbackController
 
@@ -44,6 +46,79 @@ defmodule TdAiWeb.SuggestionController do
          {:ok, %Suggestion{} = suggestion} <-
            Completion.update_suggestion(suggestion, suggestion_params) do
       render(conn, :show, suggestion: suggestion)
+    end
+  end
+
+  def availability_check(
+        conn,
+        %{
+          "resource_type" => resource_type,
+          "domain_ids" => domain_ids,
+          "template_id" => template_id
+        } = params
+      ) do
+    language = Map.get(params, "language", "en")
+    claims = conn.assigns[:current_resource]
+
+    {status, reason} =
+      with {:authorized, :ok} <-
+             {:authorized,
+              Bodyguard.permit(
+                Completion,
+                :request_suggestion,
+                claims,
+                {resource_type, domain_ids}
+              )},
+           {:template_suggestions, true} <-
+             {:template_suggestions, Templates.has_ai_suggestions(template_id)},
+           {:prompt, %{}} <-
+             {:prompt, Completion.get_prompt_by_resource_and_language(resource_type, language)} do
+        {:ok, nil}
+      else
+        {:template_suggestions, {:error, :template_not_found}} -> {:error, "template not found"}
+        {:template_suggestions, false} -> {:error, "template has no ai_suggestion fields"}
+        {:authorized, _} -> {:error, "forbidden"}
+        {:prompt, _} -> {:error, "no active prompt"}
+      end
+
+    render(conn, :availability_check, response: {status, reason})
+  end
+
+  def request(
+        conn,
+        %{
+          "resource_type" => resource_type,
+          "resource_body" => resource_body,
+          "domain_ids" => domain_ids,
+          "template_id" => template_id
+        } = params
+      ) do
+    language = Map.get(params, "language", "en")
+    %{user_id: user_id} = claims = conn.assigns[:current_resource]
+
+    with :ok <-
+           Bodyguard.permit(Completion, :request_suggestion, claims, {resource_type, domain_ids}),
+         {:ok, fields} <- Templates.suggestion_fields_for_template(template_id) do
+      FieldCompletion.resource_field_completion(
+        "business_concept",
+        resource_body,
+        fields,
+        language: language,
+        requested_by: user_id
+      )
+      |> case do
+        {:error, error} -> {:error, :unprocessable_entity, error}
+        suggestion_content -> render(conn, :show_content, suggestion_content: suggestion_content)
+      end
+    else
+      {:error, :template_not_found} ->
+        {:error, :unprocessable_entity, "template not found"}
+
+      {:error, :no_ai_suggestion_fields} ->
+        {:error, :unprocessable_entity, "template has no ai_suggestion fields"}
+
+      error ->
+        error
     end
   end