+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-6221] Secured endpoints. Only admins can access the resources.
+- [TD-6195] Created generic suggestion request and availability endpoint
 
 ## [5.19.0] 2023-11-28